Mise à jour du script de compilation pour prendre en compte les formules dans les environements speciaux.
git-svn-id: http://svn.edpsciences.org/svn/lxir/trunk/lxir@30 56c28150-6a3a-0410-800a-e81cc29300b5

+++ contrib/scripts/lxirlatex.py
@@ -1,6 +1,8 @@
 #!/bin/env python
 import os, sys, re
 
+VERSION = "0.2"
+
 translation_map = {}
 
 def find_source(name):
@@ -20,11 +22,36 @@ def temp_source_filename(source):
 		ext = '.tex'
 	filename = name + "_lxir" + ext
 	if os.path.exists(filename):
-		# raise Exception("Temporary file \"%s\" already exists" % filename)
-		os.unlink(filename)
+		if not options.overwrite:
+			raise Exception("Temporary file \"%s\" already exists" % filename)
+		else:
+			os.unlink(filename)
 	return filename
 
+def first_special_char(line, index):
+	n = -1
+	for char in ["$", "{", "}", "%"]:
+		m = line.find(char, index)
+		if m > -1 and (n == -1 or m < n):
+			n = m
+	return n
+
+def escape_math(expr):
+	expr = expr.replace("!", "!!")
+	expr = expr.replace("[", "![")
+	expr = expr.replace("]", "!]")
+	expr = expr.replace("*", "!*")
+	expr = expr.replace("-", "!-")
+	expr = expr.replace("{", "[")
+	expr = expr.replace("}", "]")
+	expr = expr.replace("^", "*")
+	expr = expr.replace("_", "-")
+	expr = expr.replace("\\", "!")
+	return expr
+
 def make_lxir_source(source):
+	if options.verbose:
+		print("processing file :%s" % source)
 	global translation_map
 	if translation_map.has_key(source):
 		return translation_map[source]
@@ -51,44 +78,69 @@ def make_lxir_source(source):
 		elif not has_requirepackage and rp.match(line):
 			has_requirepackage = True
 			d.write(line)
-		elif dc.match(line):
+		elif not has_documentclass and dc.match(line):
 			has_documentclass = True
 			if not has_requirepackage:
-				d.write("\\RequirePackage[verbatimmath]{lxir}\n")
+				if not options.verbmath:
+					d.write("\\RequirePackage{lxir}\n")
+				else:
+					d.write("\\RequirePackage[verbatimmath]{lxir}\n")
+			d.write(line)
+		elif not has_documentclass or not options.verbmath:
 			d.write(line)
 		else:
 			index = 0
+			bracket_level = 0
 			while index >= 0:
-				p1 = line.find("$", index)
-				p2 = line.find("$$", index)
-				if p2 >= 0 and (p1 < 0 or p2 <= p1):
-					if p2 > index:
-						d.write(line[index:p2])
-					if math_type == 0:
-						d.write("\n\\begin{formule}\n$$")
-						math_type = 2
-						index = p2 + 2
-					elif math_type == 2:
-						d.write("$$\n\\end{formule}")
-						math_type = 0
-						index = p2 + 2
-					elif math_type == 1:
-						d.write("$$")
-						index = p2 + 2
-					else:
-						raise Exception("Invalid math transition")
-				elif p1 >= 0:
-					if p1 > index:
-						d.write(line[index:p1])
-					if math_type == 0:
-						d.write("\n\\begin{formule}\n$")
-						math_type = 1
-					elif math_type == 1:
-						d.write("$\n\\end{formule}\n")
-						math_type = 0
-					else:
-						raise Exception("Invalid math transition")
-					index = p1 + 1
+				nindex = first_special_char(line, index)
+				if nindex > -1:
+					c = line[nindex]
+					if c == "{":
+						if math_type == 0:
+							bracket_level += 1
+							d.write(line[index:nindex + 1])
+							index = nindex + 1
+					elif c == "}":
+						if math_type == 0:
+							bracket_level -= 1
+							if bracket_level < 0:
+								raise Exception("Error in input, too many closing brackets")
+							d.write(line[index:nindex + 1])
+							index = nindex + 1
+					elif c == "%":
+						d.write(line[index:]) # keep comments (not in math ?)
+						index = -1
+					elif c == "$":
+						if math_type == 0:
+							if line[nindex:nindex+1] == "$$":
+								if bracket_level != 0:
+									raise Exception("Error in input, $$ inside brackets")
+								math_type = 2
+								d.write(line[index:nindex] + "\n\\begin{formule}\n$$")
+								index = nindex + 2
+							else:
+								if bracket_level == 0:
+									math_type = 1
+									d.write(line[index:nindex] + "\n\\begin{formule}\n$")
+									index = nindex + 1
+								else:
+									findex = line.find("$", nindex + 1)
+									if findex == -1:
+										raise Exception("Error in input, multiline math expression inside brackets")
+									d.write(line[index:nindex] + "\\verbatimmathformule{%s}" % escape_math(line[nindex + 1:findex]))
+									index = findex + 1
+						elif math_type == 1:
+								d.write(line[index:nindex] + "$\n\\end{formule}\n")
+								math_type = 0
+								index = nindex + 1
+						elif math_type == 2:
+							if line[nindex:nindex+1] == "$$":
+								d.write(line[index:nindex] + "$$\n\\end{formule}\n")
+								math_type = 0
+								index = nindex + 2
+							else:
+								d.write(line[index:nindex+1])
+								index = nindex + 1
 				else:
 					d.write(line[index:])
 					index = -1
@@ -97,6 +149,8 @@ def make_lxir_source(source):
 	return temp
 
 def compile_latex_source(filename):
+	if options.verbose:
+		print("Compiling latex source :%s" % filename)
 	result = os.system("latex -interaction=batchmode %s" % filename)
 	if result != 0:
 		raise Exception("latex compilation failed (%d)" % result)
@@ -109,16 +163,32 @@ def compile_lxir_source(dvi):
 		raise Exception("lxir translation failed (%d)" % result)
 	return xml
 
-def main():
-	try:
-		if len(sys.argv) < 2:
-			raise Exception("Usage : %s <source>\n" % sys.argv[0])
-		source = make_lxir_source(sys.argv[1])
-		dvi = compile_latex_source(source)
-		xml = compile_lxir_source(dvi)
-	except Exception, e:
-		sys.stderr.write("Error: " + str(e) + "\n")
-		sys.exit(1)
-
 if __name__ == '__main__':
+	from optparse import OptionParser
+	def main():
+		global options
+		parser = OptionParser(usage = "%prog [options] <source> [<source> ...]", version = "%prog " + VERSION)
+		parser.add_option("-m", "--verbmath", help="Use verbatim math mode", action="store_true")
+		parser.add_option("-M", "--not-verbmath", help="Don't use verbatim math mode", action="store_false", dest="verbmath")
+		parser.add_option("-v", "--verbose", help="Verbose mode", action="store_true")
+		parser.add_option("-q", "-V", "--quiet", "--not-verbose", help="Deactivate verbose mode", action="store_false", dest="verbose")
+		parser.add_option("-o", "--overwrite", help="Overwrite temporary files", action="store_true")
+		parser.add_option("-O", "--not-overwrite", help="Do not overwrite temporary files", action="store_false", dest="overwrite")
+		parser.add_option("-d", "--delete-temp", help="Delete temporary files", action="store_true")
+		parser.add_option("-D", "--not-delete-temp", help="Do not delete temporary files", action="store_false", dest="delete_temp")
+		parser.set_defaults(overwrite = True, verbmath = True, verbose = False)
+		try:
+			options, args = parser.parse_args()
+			if len(args) < 1:
+				parser.error("Need at least one source file")
+			for arg in args:
+				source = make_lxir_source(arg)
+				dvi = compile_latex_source(source)
+				xml = compile_lxir_source(dvi)
+				if options.delete_temp:
+					os.unlink(source)
+					os.unlink(dvi)
+		except Exception, e:
+			sys.stderr.write("Error: " + str(e) + "\n")
+			sys.exit(1)
 	main()
